restore FILEFORMAT
This one is a bit delicate; the default value of the parameter "literal" depends
on the particular class of inset. Therefore we have to store a fake parameter in
the params to mimic the "literal" parameter.

This is convoluted. In the future it might be helpful to implement an option to
disable the output and modification of a parameter and set it to a fixed default
value.